+++ cppm-homeworks/cppm-diplom/lib/transport/inc/Broom.h
@@ -7,7 +7,7 @@ class  Broom :
 {
 protected:
   double reduction;
-  double getDistanceReduction(int distance) const override;
+  double getDistanceReduction(double distance) const override;
 public:
   TRANSPORTRACE_API Broom();
 };
