+++ cppm-homeworks/cppm-diplom/lib/transport/inc/GndTransport.h
@@ -11,7 +11,7 @@ protected:
     virtual double calcRestTime(const int restCount) const = 0;
 public:
     TRANSPORTRACE_API GndTransport(std::string name, int speed);
-    TRANSPORTRACE_API virtual double calcTime(int distance) const;
+    TRANSPORTRACE_API virtual double calcTime(double distance) const;
     TRANSPORTRACE_API TransportType getType() const override;
 };
